Improves error handling in Base#method_missing
This fixes an issue that caused NoMethodErrors that were somewhere raised in a widget implementation to be falsely identified as NoMethodErrors raised by the view context. In case of an exception, we double check now and re-raise the exception if necessary or create a new one that indicates that the widget actually does not respond to the called method.